Extract `MockApp`.
